+++ src/net/mod.rs
@@ -1,4 +1,6 @@
+mod codec;
 mod tls;
 pub mod transport;
 
+pub use codec::CodecConfig;
 pub use tls::TlsConfig;